iio_{readdev,writedev}: Add benchmark feature
Add a new --benchmark (or -B) option.

With this option enabled, no data is written to the standard output (for
iio_readdev) or read from the standard input (for iio_writedev).
Instead, the buffers are flipped as fast as possible. This functionality
permits to benchmark the speed at which the data flows, and compare the
performance of the various backends and/or with various buffer sizes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>