Add support for hardware monitoring (hwmon) devices
Add preliminary support for hardware monitoring (hwmon) devices.

The hwmon devices can be used exactly as if they were IIO devices. The
only thing to be careful about, is that the type of the channel is one
of the hwmon_chan_type enum instead of the iio_chan_type enum.

To avoid confusion a hwmon_channel_get_type() wrapper has been added as
well, which just returns the iio_channel's type as a hwmon_chan_type.

Note that only the current (as of 2021) hwmon API is supported. The old
and deprecated APIs are not supported, and won't be supported unless we
have a case where updating a hwmon driver is not possible.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Co-developed-by: Robin Getz <robin.getz@analog.com>
Signed-off-by: Robin Getz <robin.getz@analog.com>