Update iio_device_get_sample_size()
Add an additional parameter, a pointer to a 'iio_channels_mask'. If
provided, the sample size is computed from the mask. Otherwise, it is
computed from the iio_device's internal channels mask.

There is currently no way to retrieve a pointer to a iio_channels_mask
structure, but this will be useful in the future.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>