iio.h: Add functions to handle pointer-encoded errors
These functions iio_err(), iio_ptr() and iio_err_cast() can be used to
encode error codes into pointers or retrieve an error code from a
pointer.

This will later be used by the functions of the public API, to remove
their dependency on the 'errno' variable, and therefore on the <errno.h>
specific error codes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>