scan: Rework scanning
Instead of going through the list of enabled backends and checking for
each one whether or not we were asked to scan from it, we now go through
the user-supplied backends list, find the corresponding backend, and
call its scan function.

This change will make it possible to support scanning from the same
backend multiple times, for instance with different USB devices.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>