scan: Prefer comma instead of colon for backends separator
Libiio 1.x (in the dev branch) uses a comma instead of a colon as the
backends separator in the scan list string. This makes it possible to
use the colon as the VID/PID separator in the USB backend parameters.

Update the code and documentation to specify that commas should be used,
while still allowing colons as separators to be backwards-compatible.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>