dynamic: Clean up API
Instead of using the 'struct iio_module' as an opaque pointer directly
to the underlying platform-specific type, make it a real object.

This has the benefit of providing a better API: one function to create
the iio_module, one to destroy it, and one function to get the
iio_backend from the iio_module.

The platform-specific functions are now called iio_dlopen, iio_dlclose,
and iio_dlsym, so that the old function names can be reused by the new
code.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>