scan: rework scanning feature
Since the latest merge of libiio master it was possible to specify the
list of backends to scan, but this only worked with the built-in
backends. With this commit, when a requested backend is not built-in, it
will be attempted to be loaded from the corresponding module.

Additionally, since commit 163f9a5f ("scan: Fix default backends list
containing disabled backends") we have a list of all the enabled
backends, built-in or modules; so all of our installed modules will be
scanned by default, and it makes no sense to call iio_dynamic_scan()
anymore.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>