Add new iio_block API
Introduce a new iio_block object, and enhance the iio_buffer object with
a set of new API functions.

This new API is lower-level than the previous one. If the old one was
more suited for low-speed devices that used IIO's fileio interface, the
new one is more suited for high-speed transceivers that transmit or
receive data by passing blocks of samples back and forth to the DMA
controller.

It also supports multi-buffer setups, where one iio_device has more than
one hardware iio buffer, which was impossible to handle before.

The application is now responsible for allocating the blocks of data,
and for enqueueing and dequeueing them in the correct order. It is also
responsible for enabling or disabling the buffer.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>