iiod: Enable iiod to listen on non-standard ports
This enables iiod to take a new command line argument '-p <num>' which
is port number. This effects both:
 - the port number that iiod listens on
 - the advertised port number that is used with Avahi

<num> can be any of:
             0 : uses a random/free dynamic port
    1 to  1023 : System Ports (must be root)
 1024 to 49151 : User Ports
49152 to 65535 : Dynamic Ports, or Ephemeral

This allows multiple instances of iiod to be used on the same machine,
as well as just using it on different network ports, making it easier to
bind behind ssh or something that provides security.

Signed-off-by: Robin Getz <robin.getz@analog.com>