iiod: allow iiod to also connect to USB or serial devices.
This enables iiod to be used as a "network bridge" for those devices
which are on a serial or USB backend. For example a Pluto device (USB)
which is plugged into a RPi, could now run iiod and allow the USB conext
to be "bridged" to Ethernet to other remote hosts.

Signed-off-by: Robin Getz <robin.getz@analog.com>