iio.h: Use generic macros to read/write attributes
Instead of having 36 different functions to read/write attributes, which
is unnecesarily complex, reduce it to 16: one *_read_raw(), one
*_write_raw(), one *_read() and one *_write() for each of the attribute
types (device, channel, debug, buffer).

The *_read() is a generic macro that will call the proper API function
according to the type of the value pointer. The supported pointers are
long long *, bool *, double *.

The *_write() is a generic macro that will call the proper API function
according to the type of the value. The supported value types are
const char *, char *, long long, bool, double.

While the _Generic() keyword requires C11, the whole library, utils,
examples and IIOD do not use it, and therefore can still be compiled as
C99.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>