CMake: add debug options to windows builds
gcc and clang always build with '-g', enabling easier debugging, so
catch MSVC to the same.

This will create seperate program database (.pdb) files, which map
identifiers and statements in the project's source code to
corresponding identifiers and instructions in compiled apps.
These mapping files link the debugger to your source code, which
enables debugging.

Hopefully this will put a stop to those who want to build in windows
in debug mode (which is not possible, see #772)

This does not include the pdb files in the installers, so the only way
to get them is build locally.

Signed-off-by: Robin Getz <robin.getz@analog.com>