Use FQDN_LEN instead of MAXHOSTNAMELEN for DNS names
As defined in nbase.h (or in some system header file on Linux),
MAXHOSTNAMELEN is 64. This is the maximum length of a hostname,
but not of a DNS name, which may be longer. RFC 1035 spells out
that DNS names are formed of labels, each of which must be 63
bytes or less, and that labels are combined to form a DNS name
(a.k.a. FQDN) which may not exceed 255 bytes.

So add and use that, which fixed a long name issue for me.

Signed-off-by: Robin Getz <robin.getz@analog.com>