iio.h: Update iio_channel_{read,write} functions
Update the prototype of the iio_channel_read() and iio_channel_write()
functions.

Instead of taking a pointer to an iio_buffer, they now take a pointer to
an iio_block, and will extract or insert the samples from / to the
iio_block object.

They additionally now take a "raw" boolean parameter. If set, the
samples aren't converted between the host and machine formats.

This make the iio_channel_{read,write}_raw() functions obsolete, which
have then been removed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>