iiod: Use long-lived structure for device pdata
Instead of using the device data storage to store a pointer to the
DevEntry, which is set when the device is opened and cleared when the
device is closed, set the device data pointer to a iio_device_pdata
structure, which will live as long as the context exists.

This will later allow to add more device-specific fields.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>