iiod-client: Add fall-back mechanism for .read_line callback
If the .read_line() backend function is missing, read the response codes
character by character.

The .read_line() implementations can become very complex when timeouts
are introduced, and it's just not worth it. The .read_line()
implementations for the network and serial backends will then be
removed. With the introduction of the new asynchronous interface, the
old protocol will mainly be used for very slow devices (tinyiiod)
anyway.

Only the USB implementation will remain, since it cannot read
char-by-char and can only read a line in a single operation.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>