network: Remove .read_line() implementation
The iiod-client code will use the default implementation if the backend
does not provide one.

The default implementation is simpler than what we had in the network
backend, since it reads the line character by character, while the
implementation in the network backend would read the exact number of
characters.

However, the network backend's implementation was quite complex, was
only beneficial on Linux, and with the upcoming new asynchronous API,
would be only limited to talking to legacy IIOD servers and tinyiiod
instances, in which cases we don't really care if the performance is a
bit lower.

One benefit of removing this code now is that handling timeouts in
iiod-client will be much easier now.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>