iiod-client: Use new cancel callback before destroying responder
Introduce a new .cancel() callback in the iiod_client_ops structure.

This callback will be used (if provided) before the iiod_responder is
destroyed. This fixes a deadlock problem occuring when the IIO context
fails to be created, as the iiod_responder's read thread would block
indefinitely on I/O.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>