block: Return -EPERM if trying to enqueue or dequeue twice in a row
Return a -EPERM (operation not permitted) error when trying to dequeue
or enqueue the same block several times in a row.

Trying to dequeue an already dequeued block used to be a no-op which
returned 0. However, at no point should an application try to dequeue
the same block twice, so an error should definitely be returned.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>