dynamic: Handle symbols mapped at 0x0 in iio_dlsym
When dlsym() returns a NULL pointer, check the return value of
dlerror(). If it returns non-NULL, an error occured and we can return a
pointer-encoded code. If however it does return NULL, then we found a
symbol mapped at address 0x0. In that case iio_dlsym() will return
(void *)0x0 (aka. NULL) which should not be treated as an error.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>