iiod: Use one AIO context per transfer direction
When using the new async. protocol, IIOD must be able to receive and
send at the same time. This was previously not possible when using AIO.
It didn't seem to matter with the network code, but it did matter with
the USB code.

To address this issue, use two AIO contexts, one for each transfer
direction.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>