iiod: Support zero-copy to USB
Add support for passing sample data as DMABUF objects between IIO
devices, and the USB stack.

This mechanism has the benefit that the CPU will never access the data
to copy it from one hardware buffer to another, and therefore results in
a much higher throughput at a much lower CPU usage.

For this new mechanism to work, the DMABUF-based IIO kernel API must be
available and used by the local backend, and the FunctionFS stack must
also support importing DMABUFs. If those conditions are not met, the
standard way of transferring the data will be used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>