Clarify licenses
Libiio was always meant to be LGPL, but the tools, examples and IIOD
were supposed to be GPL (as it does not really make sense for them to
use LGPL).

However, the SPDX identifiers (and the comment blocks before that) of
the source files all stated that the license was LGPL, since the
beginning.

Since re-licensing would be quite a headache, because of the huge number
of copyright holders, simply embrace the mistake and use the LGPL
license for everything.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>