Clarify licenses, round 3
Revert the previous 2 commits...

The license of the examples and test programs are indeed GPLv2+ and not
LGPLv2.1+, and the SPDX identifiers were wrong.

In the time frame where the examples and test programs were using the
incorrect SPDX identifiers, the only contributor apart from myself was
Nuno Sa, who contributed the adrv9002-iiostream.c example under the
LGPLv2+ license. He gave his consent for the relicense in the following
github thread:
https://github.com/analogdevicesinc/libiio/pull/924

Signed-off-by: Paul Cercueil <paul@crapouillou.net>