CMake: Fix cyclic dependency under MinGW
iiod-responder is compiled as a static library, because it is used both
inside Libiio, and by IIOD as well.

On Windows, it will only be compiled against Libiio, as IIOD cannot be
enabled. However, unlike a program linked against Libiio that will
import the symbols from the DLL, it is an integral part of Libiio, and
therefore should not try to import the symbol names as DLL symbols.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>