move iio-monitor from examples to tests
The application iio-monitor, was only built, if WITH_EXAMPLES was on,
but installed like a TEST application (which is weird, and complicated
some of the cmake files).

Clean that up by moving the app over to the test directory, and clean
the cmake in ./examples, and plumb it into the ./test directory.

Signed-off-by: Robin Getz <rgetz@mathworks.com>