cmake: ensure when someone turns something off, it prints out
normally when you turn something off during the cmake, you invoke it
like:
cmake ../ -DWITH_USB_BACKEND=0

it shows up internally as prefixed with a _0, not a _OFF. Today we only
print _OFF, not _0, so we silently eat things without letting people
know (which is bad form).

This patch fixes that, by making sure _1 and _ON; and _0 and _OFF are
printed out. This should make sure people know what they are building.

Signed-off-by: Robin Getz <rgetz@mathworks.com>