task.c: Fix race between thread and iio_task_sync()
After the main thread is done with an entry, and the corresponding
"done_lock" has been unlocked, the thread must not use the entry
pointer anymore, as it may have been freed in iio_task_sync(),
unless the entry is "autoclear".

In that case though, we still must not deference the entry pointer to
read the "autoclear" flag. Fix this race by caching the "autoclear"
value into a local variable, so that the entry pointer is never
referenced after the mutex is unlocked.

This fixes rare segmentation faults within IIOD when using the file-io
interface in the local backend.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>