treat warnings as errors in CI
In the past, when we build on travis and appveyor, we would detect this,
and turn on -Werror (treating warnings as errors). Somewhere along the
line - that got lost.

Re-enable that for azure pipelines, and document it in the README

At the same time, turn on the equivilent for the cmake files themselves.

Signed-off-by: Robin Getz <rgetz@mathworks.com>