packaging: update DEB_DETECT_DEPENDENCIES
when things moved from travis-ci to azure, some packaging options were
lost. Specifically DEB_DETECT_DEPENDENCIES for debian based systems.

This ensures that the version numbers from the host systems are put in
the deb files so that things are actually compatible with the
downstream systems. This patch removes the option, and makes it the default
behavior, and makes RPM based systems do the same.

basically - fixing issues that weren't seen over the last 3 years while
this was turned off.

Signed-off-by: Robin Getz <rgetz@mathworks.com>