update packaging: add requirements into debian tar.gz files
When using the tar balls - it's problematic sometime, since we don't
include the required libraries that are needed to run, and have no way to
communicate these to a end user (like a deb or rpm does).

This fixes that by adding a "required2tar" target which can be run after
"make package", which fixes this problem, by adding the required libraries
to the tar ball, just like we do for the windows zip file.

Signed-off-by: Robin Getz <rgetz@mathworks.com>