CMake: Find libusb include dir using pkg-config
Prior to this, only the library was discovered using pkg-config, while
the header was searched manually. This caused troubles on the CI when
building for OSX, as it would succeed to find the library but not the
header file.

Now, the manual search is now performed only when pkg-config cannot
locate a libusb package; which is what was already done for the library
file itself.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>