device: Fix sample size calculation
If we consider a system where channel 0 delivers 32-bit data and channel
1 delivers 16-bit data, the sample size computed would be (4 + 2) == 6
bytes.

This is not correct, as it does not take into account the padding needed
until the next sample, which must be computed in a way that all the
elements of the next sample are aligned to their own size.

This new algorithm follows what the Linux kernel does (see:
iio_compute_scan_bytes() in drivers/iio/industrialio-buffer.c as of
v6.4), and aligns the sample size to its largest element.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>