dnssd: windows: Greatly enhance code
The previous code was very complex and was not working properly with
IPv6 addresses.

The new code will use the SRV records to find the hostnames that support
IIO functionality, and then add new entries when a AAAA record provides
a new IPv6 address for a hostname we already know of.

Link-local addresses are now supported, and the interface number will be
appended to the address. Global addresses are supported as well.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>