iiod: fix the printing of IP addresses inside iiod
when a IP client attaches to the current iiod, it prints out:
New client connected from 0.0.0.0

This is because we this was coded when we only supported ipv4, but when we
got a ipv6 socket (which is all the time now when we have ipv6 enabled),
we didn't allocate the right size of structure, and the client/peer
address didn't get decoded/printed out properly.

So, fix that, and print out where the clients are at.

Tested onm x86, ARM (32-bit and 64-bit)

Signed-off-by: Robin Getz <rgetz@mathworks.com>
Signed-off-by: Paul Cercueil <paul@crapouillou.net>
(cherry picked from commit 6ab7ff64f07c0031f0779edd5a5a8cfccf2582bb)