debug.h: Update log macros
Remove the coloring feature, because the WITH_COLOR_DEBUG macro was not
set anywhere, and would trigger compilation errors when set (since we
enforce strict C99 compliance).

The log macros are also not conditionally-compiled anymore, and rely on
the compiler detecting them as dead code instead. This should fix all
issues with variables detected as unused by the compiler when the
'NoLog' level is used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>