local-mmap: Fix block allocation
The algorithm for allocating new blocks was broken, as it was requesting
an allocation of (nb_blocks + 1) blocks instead of a single block.

There is also no need to check the return value of req.count, as we are
allocating a single block now; if the allocation fails the ioctl() will
return an error code.

Fixes #1042.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>