sort: Update iio_channel_compare()
This function is not used anywhere right now, so it can be updated
without side effects.

The previous algorithm is better, but was eventually left unused as
it would change the ordering of the channels within an IIO context,
which caused communication problems between libiio and a remote IIOD.

Update iio_channel_compare() to mirror the algorithm used in
reorder_channels() in context.c.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>