utils: Update to new attributes API
Update the utility programs to use the new attributes API.

The conversion is generally straightforward, except when it comes to
buffer attributes. With the new API the buffer attributes can be
obtained from a "iio_buffer" object, which means that "iio_attr" and
"iio_info" need to create a buffer in order to list its attributes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>