iiod: Update to new attributes API
Switch to the new attributes API in the old ASCII protocol handler
(ops.c) and the new binary protocol handler (responder.c).

The conversion is mostly straightforward, except in the case of buffer
attributes with the old protocol; in that case the attributes can be
accessed only when a iio_buffer for the corresponding device has been
created beforehand.

This is not really expected to be a problem, as it would only affect
Libiio v0.x clients which make use of non-standard buffer attributes -
and that should be a really reduced population.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>