compat: Update to new attributes API
Switch to the new attributes API in the compatibility layer.

The conversion is mostly straightforward, except in the case of buffer
attributes with the old protocol; in that case the attributes can be
accessed only when a iio_buffer for the corresponding device has been
created beforehand.

This is not really expected to be a problem, as there aren't many
programs out there making use of non-standard buffer attributes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>