iiod: Move blocklist to buffer list entries
Instead of using a global lock, and a single linked list for all the
blocks created by the client, use one lock per buffer alongside one
linked list per buffer for all the blocks that were created for it.

This means that operations on the block lists of different buffers can
happen in parallel.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>