python: Update to new IIO attributes API
Use a single iio.Attr class to represent attributes, to follow the
latest developments of the Libiio v1.0 API.

This also simplifies a lot the code. The pyiio API itself almost did not
change; the only main differences are that buffer attributes are now
available from the iio.Buffer class, and there's now only a single
iio.Attr class instead of multiple ones.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>