local: Keep file descriptor of dev node opened
Keep a file descriptor to the /dev/iio:deviceX node opened. When
creating a buffer, this file descriptor will be duplicated.

This ensures that we can create a IIO events file descriptor without
having to create a IIO buffer first, and without the problem of trying
to open the main FD twice.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>