iiod-client: Add functions to support events
Add three new functions tailored for IIO events support. These functions
allow to create an IIO events stream, read from it, and close it.

The open / close functions use the regular I/O pipe of the responder,
while the event read function will use an I/O pipe dedicated to the
event stream. This allows the call to be blocking, only returning when
an event has been read, without disturbing the command flow.

Note that there is no listener thread that would continuously poll the
IIO events; those are only read when iiod_client_read_event() is called.
This is fine, as IIO events are buffered anyway on the Linux kernel
side, so events are not lost.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>