iio-private: Add padding after iio_data_format in iio_channel struct
Add padding after the iio_data_format, which will be cleared to zero
during initialization. The total of (iio_data_format + padding) will be
128 bytes which should be plenty. The point of this, is to make sure
that we can have forward compatibility with any apps built with a newer
version of Libiio that may have more fields added to the iio_data_format
structure.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>