iiod: Restore support for reading/writing all attributes
Libiio's new API does not allow reading or writing all attributes at the
same time. However, old clients connected to the old interface might
still try to use it.

Re-implement it in IIOD on top of the new API so that old clients still
work. The algorithm is based on the one in v0.25's local.c.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>