Add "cyclic" parameter to .buffer_enable() backend callback
IIOD v0.x supports enqueueing cyclic blocks, but it needs to know about
it when the iio_buffer is opened through the "OPEN" command. As
iiod-client calls this command when the buffer is to be enabled, pass an
extra "cyclic" parameter to .buffer_enable(), always set to "false" for
now.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>