iiod: Split network code into its own network.c file
Move all the network code to a new network.c. The code is slightly
different, as the network code now runs into its own thread, attached to
the main thread pool, and the client threads it will create will be
attached to a new pool.

This will later allow to disable network support at compile time.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>