iiod-responder: Allow surviving read/write issues
Instead of exiting the read thread in case there is a read error, it
will now cancel all responses to notify the issue, but keep trying to
read data.

To prevent the read thread from trying and failing reading data in a
tight loop, it will now wait on a condition variable if there are no
readers registered.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>