iiod-client: Send block dequeue request when previous one failed
The IIOD_OP_TRANSFER_BLOCK command would atomically enqueue then dequeue
the given block. The problem with that, was that on error it was not
possible to know which of the two operations was the cause of the error.

Change this by encoding the return value in a way that we can deduce
which of the two operations failed.

Then, if the enqueue failed, the block will be marked as not enqueued,
and can be enqueued again. If the dequeue failed, the block is still
marked as enqueued, and the next dequeue attempt will see the use of the
IIOD_OP_RETRY_DEQUEUE_BLOCK command, which will only dequeue the block.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>