iiod-client: Use fallback client to destroy blocks and buffers
Allow clients to pass a fallback iiod-client instance to be used when
destroying blocks and buffers.

This is useful because once a buffer is cancelled with
iio_buffer_cancel(), the iiod-client associated with that buffer cannot
be used for data transfers anymore. Using the fallback instance allows
the client to destroy the blocks and buffer properly after the buffer
has been cancelled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>