iiod: Start enqueue/dequeue tasks when creating buffer
The problem with starting the enqueue/dequeue tasks when enabling the
IIO buffer is that the blocks might not have been enqueued yet by the
time the buffer is enabled. By starting the enqueue/dequeue tasks when
creating the IIO buffer, the blocks are much more likely to have been
enqueued by the time the buffer is enabled (but still not guaranteed,
unfortunately).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>