iiod: Handle event reads in tinyiiod
Due to the fact that tinyiiod runs thread-less, it is not possible to
have a blocking read to iio_event_stream_read().

To handle event reads properly, tinyiiod based applications should
return -EAGAIN in the iio_backend_ops.read_ev() callback if there is no
event, and have the responsability to call iiod_set_event() when an
event (or error) do occur.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>