iio.h: Add IIO_MOD_LIGHT_UVA/UVB modifiers
Add the IIO_MOD_LIGHT_UVA and IIO_MOD_LIGHT_UVB modifiers that will be
introduced in Linux v6.8.

Note that Linux v6.8-rc1 added those entries in the middle of the
iio_modifier enum (which is wrong as it breaks ABI), but it will be
fixed for the regular v6.8 release.

Fixes #1130.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>