xml: Decouple from private API and use new backend functions
Use the new backend functions to create the iio_context, iio_device and
iio_channel objects. By doing so, the XML backend is now fully decoupled
from the private API, and it only uses public headers.

Note that we don't need to print the channel or device names with the
debug macros, as they already do that for us.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>