context: Handle if description is NULL and ctx->description exists
Add check when attempting to concatenate description strings.

When description was NULL and ctx->description was a valid string then
ctx->description was freed and replaced by NULL. This caused iio_info to
not print the backend description string.

Additionally, do not free(NULL) when ctx->description is NULL.

Fixes #1121

Signed-off-by: Philip Molloy <philip@philipmolloy.com>