utils: iio_common: Set timeout specified at CLI immediately
When a timeout is specified as a command-line argument then it is only
applied after the final context is created. If the creation of the
context requires transfering data then a default timeout is used. That
default timeout might be too short for slow transfers, preventing the
context from being created.

For example, a serial backend without any devices needs to send a
context XML string of roughly 1170 bytes, which when communicating at
9600 baud will take more than the default timeout of 1000 ms.

Signed-off-by: Philip Molloy <pmolloy@baylibre.com>