eliminate git sha from version when a release is created
For the PATCH value of the VERSION variable the git sha of the last
commit was used, but in case of a release we want to change that value
to 0, so it can be added in a package repository.

If we set an environment variable IS_RELEASE then the PATCH variable will
remain 0, otherwise will still be initialized with git sha.

Signed-off-by: Andreea Andrisan <Andreea.Andrisan@analog.com>