cmake: add an Utilities library
Create a cmake utilities library for functions or macros that will be
shared across CMake files. For now, just add a function to check the
availability of pthread_setname_np(3). This is in preparation of adding
support for thread names where we'll also check this from the top level
CMake file.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>