iio.h: Add C++ compiler support for generic attr
The C11 generic macros iio_attr_{read,write} cannot be accesed from a
program built with C++. Instead C++ makes use of function overload to
achieve the same behaviour.

This change does not influence the behaviour of libiio in C
environments, but it allows the use of generic attr functions in C++
programs linked to libiio.

Signed-off-by: Andrei-Fabian-Pop <Andreifabian.Pop@analog.com>