iiod: responder: properly sync free_blocks() with enqueuing
We need to completely synchronize handle_free_blocks() with operations
that might enqueue them in another task (like handle_transfer_block()).
Otherwise we can race in a way that a block is freed but it will still
be enqueued in another task which will lead to a user after free. For
doing this we add get_iio_block_unlocked() and lock around it so we can
synchronize beyond getting the block.

Note that this still allows for a block to be enqueued in a task and
then freed while in use by that task. But it does not allow for a block
to be freed __while__ being enqueued which will allow for a future fix where
we can properly sync tokens before freeing the block.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>