task: allow to control token lifetime in iio_task_sync()
By default, the struct iio_task_token token is destroyed after syncing the
task. In preparation for adding tokens with no bounded lifetimes
(meaning that outer callers can create and destroy tokens as needed),
split iio_task_sync() so we can control if the token is automatically
destroyed or not.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>