task: add iio_task_token_do_enqueue()
This change is also in preparation for adding support for creating and
destroying struct iio_task_token tokens with full control of their
lifetimes. Hence, iio_task_token_do_enqueue() is added so that we can
reuse the code that adds a token to a task (and better control if it's
an existing vs a new token).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>