task: add API to fully control struct iio_task_token
This adds APIs to create, enqueue and destroy (destroying just turns the
already existing function public)  struct iio_task_token tokens. This
will be important in fixing a race between freeing IIO blocks and
enqueue/dequeueing them since with this, we will be able to properly sync
on the tokens.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>