iiod: responder: fix race with IIO blocks
IIOD was often crashing during client removal when buffering. The
reason was that blocks could be enqueued in the (d)enqueue tasks, then
freed and only then the tasks would process the already freed block. In
order to fix this (and now that we have no possible race between
handle_free_block() and handle_transfer_block()), we need to sync on the
(d)enqueue block tokens before freeing it. Like this we can be
sure that the block won't be used after it is freed.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>