csharp: Ensure iio scan blocks get destroyed after use
Avoid keeping a pointer to a scan block that was never destroyed, either
at the end of the ScanContext object's lifecycle or when calling any of
the 'get_dns_sd_backend_contexts', 'get_usb_backend_contexts', or
'get_local_backend_contexts' functions. These functions replace the
pointer with a new one, leaving the old pointer undestroyed.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>