stream: dequeue blocks on stream destroy
Make sure to dequeue any possible in flight block before destroying it.
Note that if the block is not enqueued, this returns an error but we
really don't care about it. Making sure the block is dequeued makes
for a much cleaner exit specially in cases where freeing the blocks is
deferred to the default client.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>