device.c: Avoid reordering channels for XML based iio contexts
The XML already contains channels in a sorted order.

These changes reduce context creation time, but the primary reason for
implementing them is to address differences in channel sorting behavior
between Windows and Linux. These differences cause communication issues
between libiio running on Windows and the remote IIOD.

More specifically, when all channels lack an index (marked with a value
of -2), qsort() on Linux preserves the original order since there is
nothing to sort. However, on Windows, the channel list is modified each
time qsort() is called.

The channel's position in the list is critical because it is used by the
binary interface with IIOD.

The first symptom of this issue was the failure to create an
iio_context. Internally, this process reads all scale attributes from
all channels of all devices. One of these read operations failed because
the command contained the channel position, which was altered, making it
unrecognizable by IIOD.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>